:sparkles: Properly display tool call (all steps)
